+++ CardGame/Model/CardMatchingGame.m
@@ -86,7 +86,7 @@ static const int COST_TO_CHOOSE = 1;
   return (index < [self.cards count]) ? self.cards[index] : nil;
 }
 
-- (void)removeCardFromIndex:(int)index {
+- (void)removeCardFromIndex:(NSUInteger)index {
   [self.cards removeObjectAtIndex:index];
 }
